chore: added NO_CLEANUP env for smoke tests (unrelated)        use exception2Result

+++ smoke/esm/it.sh
@@ -75,6 +75,14 @@ EOF
 
 pnpm install
 pnpm run test
-rm -rf $tmpDir
-cd $smokeDir
+
+if [ -z "$NO_CLEANUP" ]
+then
+  rm -rf $tmpDir
+else
+  echo $tmpDir
+fi
+#rm -rf $tmpDir
+#cd $smokeDir
 #$dockerCompose down
+
